fz.plaid_battery
